backend/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
-e git+https://github.com/rense/django-rest-camel#egg=django-rest-camel
argon2-cffi==18.1.0
certifi==2018.4.16        # via requests
cffi==1.11.5              # via argon2-cffi
chardet==3.0.4            # via requests
defusedxml==0.5.0         # via python3-openid
django-allauth==0.36.0    # via django-rest-auth
django-anymail[mailgun]==3.0
django-cors-headers==2.4.0
django-extensions==2.1.0
django-rest-auth[with_social]==0.9.3
django==2.1
djangorestframework-jwt==1.11.0
djangorestframework==3.8.2
idna==2.7                 # via requests
invoke==1.1.1
mysqlclient==1.3.13
oauthlib==2.1.0           # via requests-oauthlib
pillow==5.2.0
pycparser==2.18           # via cffi
pyjwt==1.6.4              # via djangorestframework-jwt
python3-openid==3.1.0     # via django-allauth
pytz==2018.5              # via django
requests-oauthlib==1.0.0  # via django-allauth
requests==2.19.1          # via django-allauth, django-anymail, requests-oauthlib
six==1.11.0               # via argon2-cffi, django-anymail, django-extensions, django-rest-auth
urllib3==1.23             # via requests
werkzeug==0.14.1
